Добавлена анимация удара + логика разных видов заклинаний + время хода + немного рефакторинга

+++ Assets/Scripts/GameLogic/DataModels/OffensiveSpell.cs
@@ -2,7 +2,7 @@
 {
     public class OffensiveSpell : Spell
     {
-        public int Damage { get; }
+        public double Damage { get; }
         public int FireRate { get; }
 
         public OffensiveSpell(Element majorElement, int level, string name, int damage, int mana, int fireRate, Element minorElement = default)
